+++ R/qb_extras.R
@@ -18,14 +18,9 @@
 #' }
 summarize_app <- function(subdomain, auth, app_id, agent = NULL){
 
-  if(!stringr::str_detect(auth, "^QB-USER-TOKEN ") &
-     !stringr::str_detect(auth, "^QB-TEMP-TOKEN ")){
-    auth <- stringr::str_c("QB-USER-TOKEN ", auth)
-  }
+  stopifnot(val_subdomain(subdomain), is.character(app_id), length(app_id) == 1)
 
-  if(!stringr::str_detect(subdomain, "\\.+")){
-    subdomain <- stringr::str_c(subdomain, ".quickbase.com")
-  }
+  auth <- val_token(auth)
 
   app <- qbr::get_app(subdomain, auth, app_id, agent, T, T)
   users <- qbr::get_users(subdomain, auth, agent, app_ids = list(app_id))