+++ man-roxygen/skip.R
@@ -1,0 +1,2 @@
+#' @param skip Optional. Integer. The number of rows to skip from the top of a
+#'   record set.