+++ man-roxygen/table_id.R
@@ -1,2 +1,2 @@
 #' @param table_id Character vector with one element. Found in the URL of a
-#'   Quickbase table between /db/ and ?
+#'   Quickbase table.