+++ man-roxygen/top.R
@@ -1,0 +1,2 @@
+#' @param top Optional. Integer. The limit on the number of records to pull
+#'   starting at the top of a record set.